refactor: split epever/configurer.go by concern [sctl-lyw]
Pure file moves: battery profile, charging parameters, controller time,
and config caching each get their own file, matching the existing test
split. Shared plumbing (getConfig, writeSingle, voltage validation and
block write) stays in configurer.go.
